pyproject.toml
[project]
name = "ChatX"
version = "0.1.0"
description = "Python project template 🐍"
authors = [
    {name = "Macktireh", email = "abdimack97@gmail.com"},
]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "alembic>=1.17.1",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "sse-starlette>=3.0.3",
    "uvicorn[standard]>=0.38.0",
]
requires-python = ">=3.14"

[dependency-groups]
lint = [
    "ruff>=0.14.3",
]

[tool.pdm.scripts]
# dev = "python start.py"
dev = "uvicorn main:app --reload --host localhost --port 8000"
makemigrations = "alembic revision --autogenerate"
migrate = "alembic upgrade head"
post_init = "pdm run setup"

[tool.pdm]
distribution = false

[tool.ruff]
line-length = 99
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
